kirana-copilot-ai-main/backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
transformers==4.36.0
torch==2.1.0
pillow==10.1.0
numpy==1.24.3
opencv-python==4.8.1.78
easyocr==1.7.1
pytesseract==0.3.10
pandas==2.1.4
